+++ MiniApp/Classes/JavascriptBridge/MiniAppJavaScriptEnumerations.swift
@@ -3,6 +3,7 @@ enum MiniAppJSActionCommand: String {
     case getCurrentPosition
     case requestPermission
     case requestCustomPermissions
+    case shareInfo
 }
 
 enum JavaScriptExecResult: String {
@@ -10,14 +11,9 @@ enum JavaScriptExecResult: String {
     case onError
 }
 
-enum MiniAppJavaScriptError: String {
-    case internalError
-    case unexpectedMessageFormat
-    case invalidPermissionType
-}
-
 enum MiniAppSupportedSchemes: String {
-    case tel
+    case tel    // used for phone calls
+    case about  // used to reveal internal state and built-in functions (e.g.: alert dialog)
 }
 
 /// List of Device Permissions supported by the SDK that can be requested by a Mini app
@@ -47,26 +43,13 @@ public enum MiniAppCustomPermissionGrantedStatus: String, Codable {
     case allowed = "ALLOWED"
     case denied = "DENIED"
     case permissionNotAvailable = "PERMISSION_NOT_AVAILABLE"
-}
 
-public enum MiniAppPermissionResult: Error {
-    /// User has explicitly denied authorization
-    case denied
-    /// User has not yet made a choice
-    case notDetermined
-    /// Host app is not authorized to use location services
-    case restricted
-}
-
-extension MiniAppPermissionResult: LocalizedError {
-    public var errorDescription: String? {
+    public var boolValue: Bool {
         switch self {
-        case .denied:
-            return NSLocalizedString("Denied", comment: "Permission Error")
-        case .notDetermined:
-            return NSLocalizedString("NotDetermined", comment: "Permission Error")
-        case .restricted:
-            return NSLocalizedString("Restricted", comment: "Permission Error")
+        case .allowed:
+        return true
+        default:
+        return false
         }
     }
 }