+++ MiniApp/Classes/JavascriptBridge/MiniAppJavaScriptPermissionManager.swift
@@ -3,14 +3,14 @@ extension MiniAppScriptMessageHandler {
         cachedUnknownCustomPermissionRequest.removeAll()
         userAlreadyRespondedRequestList.removeAll()
         guard let requestParamValue = requestParam?.permissions else {
-            executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppCustomPermissionError(customPermissionError: .invalidCustomPermissionRequest))
+            executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppErrorMessage(MASDKCustomPermissionError.invalidCustomPermissionRequest))
             return
         }
 
         guard let miniAppPermissionRequestModelList = prepareCustomPermissionsRequestModelList(
             permissionList: requestParamValue),
             miniAppPermissionRequestModelList.count > 0 else {
-            executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppCustomPermissionError(customPermissionError: .invalidCustomPermissionsList))
+            executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppErrorMessage(MASDKCustomPermissionError.invalidCustomPermissionsList))
             return
         }
         checkCustomPermissionsRequestStatusInCache(miniAppPermissionRequestModelList: miniAppPermissionRequestModelList, callbackId: callbackId)
@@ -48,7 +48,7 @@ extension MiniAppScriptMessageHandler {
                 cachedUnknownCustomPermissionRequest.append(
                     MiniAppCustomPermissionsListResponse(
                         name: $0.name ?? "UNKNOWN_REQUEST",
-                        isGranted:
+                        status:
                         MiniAppCustomPermissionGrantedStatus.permissionNotAvailable.rawValue))
                 return
             }
@@ -68,11 +68,7 @@ extension MiniAppScriptMessageHandler {
                 self.miniAppKeyStore.storeCustomPermissions(permissions: result, forMiniApp: self.miniAppId)
                 self.sendCustomPermissionsJsonResponse(result: result, callbackId: callbackId)
            case .failure(let error):
-               if !error.localizedDescription.isEmpty {
-                   self.executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: error.localizedDescription)
-                   return
-               }
-               self.executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppCustomPermissionError(customPermissionError: .hostAppError))
+                self.executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppErrorMessage(error))
            }
         }
     }
@@ -80,7 +76,7 @@ extension MiniAppScriptMessageHandler {
     func sendCachedSuccessResponse(result: [MASDKCustomPermissionModel], callbackId: String) {
         var permissionListResponse = [MiniAppCustomPermissionsListResponse]()
         result.forEach {
-            permissionListResponse.append(MiniAppCustomPermissionsListResponse(name: $0.permissionName.rawValue, isGranted: $0.isPermissionGranted.rawValue))
+            permissionListResponse.append(MiniAppCustomPermissionsListResponse(name: $0.permissionName.rawValue, status: $0.isPermissionGranted.rawValue))
         }
         cachedUnknownCustomPermissionRequest.forEach {
             permissionListResponse.append($0)
@@ -93,10 +89,7 @@ extension MiniAppScriptMessageHandler {
     }
 
     func sendSuccessResponse(result: [MiniAppCustomPermissionsListResponse], callbackId: String) {
-        guard let responseString = getJsonSuccessResponse(result: result) else {
-            self.executeJavaScriptCallback(responseStatus: .onError, messageId: callbackId, response: getMiniAppCustomPermissionError(customPermissionError: .unknownError))
-            return
-        }
+        let responseString = getJsonSuccessResponse(result: result)
         self.executeJavaScriptCallback(responseStatus: .onSuccess, messageId: callbackId, response: responseString)
     }
 
@@ -109,10 +102,10 @@ extension MiniAppScriptMessageHandler {
     func retrieveAllPermissionsMiniAppRequested(result: [MASDKCustomPermissionModel]) -> [MiniAppCustomPermissionsListResponse] {
         var permissionListResponse = [MiniAppCustomPermissionsListResponse]()
         result.forEach {
-            permissionListResponse.append(MiniAppCustomPermissionsListResponse(name: $0.permissionName.rawValue, isGranted: $0.isPermissionGranted.rawValue))
+            permissionListResponse.append(MiniAppCustomPermissionsListResponse(name: $0.permissionName.rawValue, status: $0.isPermissionGranted.rawValue))
         }
         userAlreadyRespondedRequestList.forEach {
-            permissionListResponse.append(MiniAppCustomPermissionsListResponse(name: $0.permissionName.rawValue, isGranted: $0.isPermissionGranted.rawValue))
+            permissionListResponse.append(MiniAppCustomPermissionsListResponse(name: $0.permissionName.rawValue, status: $0.isPermissionGranted.rawValue))
         }
         cachedUnknownCustomPermissionRequest.forEach {
             permissionListResponse.append($0)
@@ -121,9 +114,9 @@ extension MiniAppScriptMessageHandler {
     }
 
     /// For a given [MiniAppCustomPermissionsListResponse], this method will use JSONEncoder to encode the [objects] to JSON string
-    /// - Parameter result: [MiniAppCustomPermissionsListResponse] that contains the name and isGranted status of every permission that will be sent back to the Mini app as JSON
+    /// - Parameter result: [MiniAppCustomPermissionsListResponse] that contains the name and status of every permission that will be sent back to the Mini app as JSON
     /// - Returns: JSON Response string
-    func getJsonSuccessResponse(result: [MiniAppCustomPermissionsListResponse]) -> String? {
+    func getJsonSuccessResponse(result: [MiniAppCustomPermissionsListResponse]) -> String {
         let responseObject = MiniAppCustomPermissionsResponse(permissions: result)
         do {
             let jsonData = try JSONEncoder().encode(responseObject)