+++ MiniApp/Classes/Protocols/MiniAppMessageModels.swift
@@ -1,0 +1,3 @@
+public struct MiniAppShareContent {
+    public var messageContent: String
+}