+++ MiniApp/Classes/Utilities/UrlParser.swift
@@ -3,7 +3,6 @@
  */
 struct UrlParser {
 
-    static let separatorKey: String = "map-published-v2/"
     static let dropLimit = 2
 
     /**
@@ -11,14 +10,14 @@ struct UrlParser {
      * @param { url: String } - URL to parse through.
      * @return { String } - The file directory after removing the base URL.
      */
-    static func getFileStoragePath(from url: String) -> String? {
-        if url.contains(separatorKey) {
-            var pathComponents: [String] = url.components(separatedBy: separatorKey)
-            guard let path: String = pathComponents.last, !path.isEmpty else {
+    static func getFileStoragePath(from url: String, with environmemt: Environment) -> String? {
+        if let baseUrl = environmemt.baseUrl {
+            let path: String = url.replacingOccurrences(of: baseUrl.absoluteString, with: "")
+            if path.isEmpty {
                 MiniAppLogger.e("MiniAppSDK: Failed parsing URL.")
                 return nil
             }
-            pathComponents = path.components(separatedBy: "/")
+            let pathComponents = path.components(separatedBy: "/")
             // Number of elements to drop from the after the separator
             // For eg., /min-abc/ver-abc/img/test.png
             // Following code will remove /min-abc/ver-abc/