+++ Tests/Unit/Helpers.swift
@@ -9,6 +9,19 @@ class MockAPIClient: MiniAppClient {
     var zipFile: String?
     var headers: [String: String]?
 
+    init() {
+        let bundle = MockBundle()
+        super.init(with:
+                MiniAppSdkConfig(
+                    baseUrl: bundle.mockEndpoint,
+                    rasAppId: bundle.mockAppId,
+                    subscriptionKey: bundle.mockSubscriptionKey,
+                    hostAppVersion: bundle.mockHostAppUserAgentInfo,
+                    isTestMode: bundle.mockTestMode
+                )
+        )
+    }
+
     override func getMiniAppsList(completionHandler: @escaping (Result<ResponseData, Error>) -> Void) {
         guard let urlRequest = self.listingApi.createURLRequest() else {
             return completionHandler(.failure(NSError.invalidURLError()))
@@ -156,7 +169,7 @@ class MockBundle: EnvironmentProtocol {
     var mockAppId: String?
     var mockSubscriptionKey: String?
     var mockAppVersion: String?
-    var mockEndpoint: String?
+    var mockEndpoint: String? = "https://www.example.com/"
     var mockTestMode: Bool?
     var mockHostAppUserAgentInfo: String?
 
@@ -204,7 +217,9 @@ class MockMessageInterface: MiniAppMessageProtocol {
     var mockUniqueId: Bool = false
     var locationAllowed: Bool = false
     var customPermissions: Bool = false
-    var permissionError: Error?
+    var permissionError: MASDKPermissionError?
+    var customPermissionError: MASDKCustomPermissionError?
+    var messageContentAllowed: Bool = false
 
     func getUniqueId() -> String {
         if mockUniqueId {
@@ -217,27 +232,35 @@ class MockMessageInterface: MiniAppMessageProtocol {
         }
     }
 
-    func requestPermission(permissionType: MiniAppPermissionType, completionHandler: @escaping (Result<String, Error>) -> Void) {
+    func requestPermission(permissionType: MiniAppPermissionType, completionHandler: @escaping (Result<MASDKPermissionResponse, MASDKPermissionError>) -> Void) {
         if locationAllowed {
-            completionHandler(.success("Allowed"))
+            completionHandler(.success(.allowed))
         } else {
             if permissionError != nil {
                 completionHandler(.failure(permissionError!))
                 return
             }
-            completionHandler(.failure(MiniAppPermissionResult.denied))
+            completionHandler(.failure(.denied))
         }
     }
 
-    func requestCustomPermissions(permissions: [MASDKCustomPermissionModel], completionHandler: @escaping (Result<[MASDKCustomPermissionModel], Error>) -> Void) {
+    func requestCustomPermissions(permissions: [MASDKCustomPermissionModel], completionHandler: @escaping (Result<[MASDKCustomPermissionModel], MASDKCustomPermissionError>) -> Void) {
         if customPermissions {
             completionHandler(.success(permissions))
         } else {
-            if permissionError != nil {
-                completionHandler(.failure(permissionError!))
+            if customPermissionError != nil {
+                completionHandler(.failure(customPermissionError!))
                 return
             }
-            completionHandler(.failure(MiniAppPermissionResult.restricted))
+            completionHandler(.failure(.unknownError))
+        }
+    }
+
+    func shareContent(info: MiniAppShareContent, completionHandler: @escaping (Result<MASDKProtocolResponse, Error>) -> Void) {
+        if messageContentAllowed {
+            completionHandler(.success(.success))
+        } else {
+            completionHandler(.failure(NSError(domain: "ShareContentError", code: 0, userInfo: nil)))
         }
     }
 }
@@ -284,6 +307,10 @@ class MockMiniAppCallbackProtocol: MiniAppCallbackProtocol {
 }
 
 class MockNavigationView: UIView, MiniAppNavigationDelegate {
+    func miniAppNavigation(shouldOpen url: URL, with externalLinkResponseHandler: @escaping MiniAppNavigationResponseHandler) {
+        externalLinkResponseHandler(url)
+    }
+
     weak var delegate: MiniAppNavigationBarDelegate?
     var hasReceivedBack: Bool = false
     var hasReceivedForward: Bool = true
@@ -340,20 +367,12 @@ func deleteStatusPreferences() {
     UserDefaults.standard.removePersistentDomain(forName: "com.rakuten.tech.mobile.miniapp")
 }
 
-func tapAlertButton(title: String, actions: [UIAlertAction]?) {
-    typealias AlertHandler = @convention(block) (UIAlertAction) -> Void
-
-    guard let action = actions?.first(where: {$0.title == title}), let block = action.value(forKey: "handler") else { return }
-    let handler = unsafeBitCast(block as AnyObject, to: AlertHandler.self)
-    handler(action)
-}
-
 func decodeMiniAppError(message: String?) -> MiniAppErrorDetail? {
     guard let errorData = message?.data(using: .utf8) else {
         return nil
     }
-    guard let errorMessage = ResponseDecoder.decode(decodeType: MiniAppError.self, data: errorData)  else {
+    guard let errorMessage = ResponseDecoder.decode(decodeType: MiniAppErrorDetail.self, data: errorData) else {
         return nil
     }
-    return errorMessage.error
+    return errorMessage
 }